requirements.txt
fastapi
pdfplumber
openai
pandas
python-dotenv
uvicorn
